refactor push and pull CLI commands
